........
premature opt
........
7.2* resolve some hints
........
7.2* apply patch of Fabio Luis -> Ticket #330
........
7.2* wrong result for UnicodeToInt
........
7.2* apply patch for Ticket #270
........
close mariadb memleak (array bindings)

........
7.2* apply fpc hint removals thx
........
7.2* uncommend the retained commits of FB
........
add missing descriptions
........


git-svn-id: http://svn.code.sf.net/p/zeoslib/code-0/trunk@5476 d681b350-854c-439c-a81d-c0ce20710203

+++ src/core/ZFastCode.pas
@@ -341,7 +341,7 @@ function RawToUInt64Def(const S: RawByteString; const Default: UInt64) : UInt64;
 
 function UnicodeToIntDef(const S: ZWideString; const Default: Integer) : Integer; overload;
 function UnicodeToIntDef(const S: PWideChar; const Default: Integer) : Integer; overload;
-function UnicodeToIntDef(Buf, PEnd: PWideChar; Default: Integer) : UInt64; overload;
+function UnicodeToIntDef(Buf, PEnd: PWideChar; Default: Integer) : Integer; overload;
 function UnicodeToInt64Def(const S: ZWideString; const Default: Int64) : Int64; overload;
 function UnicodeToInt64Def(const S: PWideChar; const Default: Int64) : Int64; overload;
 function UnicodeToInt64Def(Buf, PEnd: PWideChar; Default: Integer) : Int64; overload;
@@ -4218,7 +4218,7 @@ begin
       Result := Default;
 end;
 
-function UnicodeToIntDef(Buf, PEnd: PWideChar; Default: Integer) : UInt64;
+function UnicodeToIntDef(Buf, PEnd: PWideChar; Default: Integer) : Integer;
 var P: PWideChar;
 begin
   P := PEnd;
